// server/types.ts
export type CommentType = 'question' | 'answer';
// ...
export interface Comment {
  // TODO: Task 2 - Create the `Comment` interface
  _id?: ObjectId;
  text: string;
  commentBy: string;
  commentDateTime: Date; 
}

/**
 * Interface extending the request body when adding a comment to a question or an answer, which contains:
 * - id - The unique identifier of the question or answer being commented on.
 * - type - The type of the comment, either 'question' or 'answer'.
 * - comment - The comment being added.
 */
export interface AddCommentRequest extends Request {
  body: {
    // TODO: Task 2 - Add fields
    _id: string;
    type: CommentType;
    comment: Comment
  };
}